+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Fix for the daily log file name, when the plugin save both log files.
     * Removed ApplePay data from the WebSDK Payment Methods, before pass them to the front-end.
     * When for SimplyConnect pay buttons is set option "text", the plugin will not subscribe for total amount change event.
+    * When the plugin can not find an Order by DMN data, do not try to create Order, but return 400 to the Cashier.
 ```
 
 # 3.1.0